Inliner fraser som ikke er gjenbrukte.
